// src/models/OtherServices.model.ts
export type OtherServicesListResponse = {
  _id: string
  serviceName: string
  serviceDescription: string
  staticPageId: string
  isActive: boolean
  isDeleted: boolean
  createdAt: string
  updatedAt: string
  __v: number
  staticPageName: string
  staticPageUrl: string
  staticPageContent: string
};